// src/Helpers/Enums/CustomerMessageEnum.ts
export enum CustomerMessageEnum {
  success = 'Succeded',
  deleted = 'Customer Successfully Deleted',
  updated = 'Customer Successfully Updated',
  created = 'Customer Successfully Created',
  failure = 'Customer Unknown Failure',
  NoDataFound = 'No Data Found',
  CustomerFirstNameRequired = 'Customer First Name Is Required',
}